test(rendering): colocate the render-context suite beside its source
`tests/rendering/render-context.test.ts` and
`src/rendering/context/render-context.test.ts` both covered the same module.
Merge the former into the latter so the suite sits next to its source, as the
repository guide requires, and so the Node and Bun runners see it at all: their
patterns select `src/`, not `tests/`.

Every case is preserved. `cspUserHeader` is dropped from the two handler-context
fixtures because `HandlerContext` has no such property; the moved file was never
typechecked in its old location and nothing read the value.

Refs veryfront/veryfront-issue-inbox#555
